加密

+++ gateway/src/main/java/com/slg/module/connection/ServerChannelManage.java
@@ -12,20 +12,20 @@ import java.util.concurrent.ConcurrentHashMap;
 @Component
 public class ServerChannelManage {
     //目标服务器连接管理
-    private Map<Integer, Channel> serverChannelMap = new ConcurrentHashMap<>();//<serverId,channel>
-    private Map<Integer, ServerConfig> serverConfigMap = new ConcurrentHashMap<>();////<serverId,ServerConfig>
+    private Map<Byte, Channel> serverChannelMap = new ConcurrentHashMap<>();//<serverId,channel>
+    private Map<Byte, ServerConfig> serverConfigMap = new ConcurrentHashMap<>();////<serverId,ServerConfig>
 
-    public Channel getChanelByIp(Integer serverId) {
+    public Channel getChanelByIp(Byte serverId) {
         return serverChannelMap.getOrDefault(serverId, null);
     }
 
-    public Channel removeChanelByIp(Integer serverId) {
+    public Channel removeChanelByIp(Byte serverId) {
         Channel remove = serverChannelMap.remove(serverId);
         serverConfigMap.remove(serverId);
         return remove;
     }
 
-    public void put(Integer serverId, Channel channel, ServerConfig serverConfig) {
+    public void put(Byte serverId, Channel channel, ServerConfig serverConfig) {
         serverChannelMap.put(serverId, channel);
         serverConfigMap.put(serverId, new ServerConfig(serverId, serverConfig.getIp(), serverConfig.getPort()));
     }
@@ -33,7 +33,7 @@ public class ServerChannelManage {
     public ServerChannelManage() {
     }
 
-    public Map<Integer, ServerConfig> getServerConfigMap() {
+    public Map<Byte, ServerConfig> getServerConfigMap() {
         return serverConfigMap;
     }
 }
